Moved LogicState to it's own file
